Add support for relative-days
Renders days ±a week as "last <day>" / "next <day>", and if it's
too long ago as -<num> for days in the past and +<num> for days
in the future.

Used for a production system where the client wants to quickly see
how far in the future this order is, or how far behind it's lagging.